Add global classmethod `from_function` to PiscesMeta.
The from_function constructor is an entry-point for user-defined
functions and data to be converted into row instances.  The user
supplies a function that accepts *args and **kwargs, and returns a list
of dictionaries appropriate to that class.  The dictionaries are
unpacked into the default class constructor, and returned in a list.