Plugins for sac2db.py, dbout is now optional --prefix.
Almost all processing is custom, so adding plugins to sac2db to extend
the default behavior is necessary.  A --plugin flag is added, which
allows users to point to a module and a function.  The flag looks like
"--plugin path/to/module_file:function_name".  The function must have a
calling signature like "def function_name(**rows):", where "rows" is a
dictionary of canonical table name keys, and corresponding values are
lists of related table class instances.  This input is precisely the
output from sac2db.dicts2rows.  Each plugin is imported and executed in
the order in which it was specified, and all plugins are applied after
the sac2db.make_atomic function (which mostly updates ids).

Also, the old "dbout" required positional argument is moved to the
optional "--prefix" argument.  Thus, standard table names can be
specified without needing a '-' dbout value; they're just the default.