Begin re-org of io/sac.py.
The goal is to break down all header-css conversions into smaller
pieces, either functions or decorators or generators, and to chain them
together somehow in order to compose more complicated functions.  The
smaller functions are then reusable between header-css conversions, and
hopefully the composite functions are more readable and configurable.