Now installs on Windows
The command 'python setup.py install' now produces no errors when
building on Windows. The libraries libecompression and libconvert can be
loaded by a ctypes.CDLL call and the functions are callable. Python at
least says there is a function pointer present.

Integrity of method calls still need to be tested for errors and valid
input/output relation. It is likely that the python methods that rely on
libecompression and libconvert are not working properly. All other
python functionality seems stable, as expected.

More preprocessor directives can be added to make a single package to
successfully on Unix and Windows once capabilities have been verified.