Use Sphinx for docs
We use Sphinx, MyST, and sphinx-autodoc2.  This makes it possible to use
the Sphinx ecosystem, which allows markdown in docs now.  We also use
autodoc2 for automatic API documentation, which doesn't yet support
numpydoc (yet?). Since our docstrings are numpydoc, we have to interpret
docstrings as markdown, which leads to some ugly interpretations.  My
hope is that numpydoc support comes soon.