Sphinx docs and GitHub Actions (#60)
* Move to src directory

* Clean Client docstring

* Move docs root (again)

* Ignore generated API docs

* Fix indentation

* Use Sphinx for docs

We use Sphinx, MyST, and sphinx-autodoc2.  This makes it possible to use
the Sphinx ecosystem, which allows markdown in docs now.  We also use
autodoc2 for automatic API documentation, which doesn't yet support
numpydoc (yet?). Since our docstrings are numpydoc, we have to interpret
docstrings as markdown, which leads to some ugly interpretations.  My
hope is that numpydoc support comes soon.

* Add TOC, split index page content

* Make a GitHub Action for docs

Docs are deployed to a blank gh-pages branch when manually triggered or on a
release.

* Fix PDF download link

* Change from autodoc2 to apidoc

Autodoc2 is great, but it doesn't understand numpydoc docstrings yet.
 I'm just commenting out the config so it can be added in easily later.

* Fix apidoc headings

User our own templates for package and module headings.
The default ones are wordy.
Also, add docstrings where they are missing.

* Append Click CLI documentation

* Remove aspirational files

* Add linked changelog

* Silence many build warnings

* Add favicon

* Remove unused env file

* Remove old LANL logo

* Admit quickstart is actually installation