String expressions (#66)
* Add flexible string filtering functions

Filtering an ORM string column (e.g. Sitechan.chan) with strings like
'BHZ', 'BH?', or ['BHZ', 'BHN'] can be complicated.  One can produce OR
and LIKE clauses that handles most of these, but this is imprecise and
may not be handled well by SQL query optimizers.  This function produces
a statement, appropriate for the input, that can be used as a query filter.